move public_values tests
